fix: Restore green CI on agent-main [ORB-10434]
Agent-Model: claude-opus-5
